+++ scalate-core/src/main/scala/org/fusesource/scalate/support/ScalaCompiler.scala
@@ -19,7 +19,7 @@ package org.fusesource.scalate.support
 
 import org.fusesource.scalate._
 import osgi.{BundleHeaders, BundleClassPathBuilder, BundleClassLoader}
-import scala.tools.nsc.Global
+import scala.tools.nsc.interactive.Global
 import scala.tools.nsc.Settings
 import scala.tools.nsc.io.AbstractFile
 import scala.tools.nsc.reporters.ConsoleReporter